requirements.txt
faiss-cpu
sentence-transformers
pypdf
langchain
langchain-community
ollama
streamlit
